samples: Bluetooth: Refactor observer for bsim test reuse
Refactor the observer sample so that it can be reused as
bsim test to verify Extended advertisement chaining.
