+++ samples/bluetooth/observer/CMakeLists.txt
@@ -4,7 +4,7 @@ cmake_minimum_required(VERSION 3.20.0)
 find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
 project(observer)
 
-FILE(GLOB app_sources src/*.c)
-target_sources(app PRIVATE ${app_sources})
-
-zephyr_library_include_directories(${ZEPHYR_BASE}/samples/bluetooth)
+target_sources(app PRIVATE
+  src/main.c
+  src/observer.c
+)
